BitBox02 Electrum plugin support
This commit adds support for the BitBox02 hardware wallet.
It supports both single and multisig for the electrum gui wallet.

To use the plugin a local installation of the BitBox02 python library is
required. It can be found on PiPy under the name 'bitbox02' and can be
installed from the bitbox02-firmware repository in the py/bitbox02
directory.

All communication to and from the BitBox02 is noise encrypted, the keys
required for this are stored in the wallet config file under the
bitbox02 key.

The BitBox02 registers a multisig configuration before allowing
transaction signing. This multisig configuration includes the threshold,
cosigner xpubs, keypath, a variable to indicate for mainnet and testnet,
and a name that the user can choose during configuration registration.
The user is asked to register the multisig configuration either during
address verification or during transaction signing.

The check the xpub of the BitBox02 for other hardware wallets, a button
is added in the wallet info dialog.

Since the BitBox02 does some things differently compared to other
hardware wallets, some special cases had to be added. These should be
removed/refactored after some review, or a later point in time. These
include:

1. The BitBox02 does not accept arbitrary keypaths to retrieve a xpub.
   Thus the password derivation was changed to be the:
   `hmac(xpub at m/84'/1'/0' || b"electrum_password_lulz")`
   To facilitate this a special case had to be added to base_wizard.py,
   where the BitBox02 is distinguished by its label. This is also a bad
   comparison, because its stringy and the field is not necessarily
   meant for this purpose.

2. The BitBox02 has a second interface for u2f. This is filtered in a
   special case in plugin.py

3. The BitBox02's hid id contains its version, not a unique id for each
   device. To still be able to uniquely identify different BitBox02's
   the id is set to the BitBox02's hid path in plugin.py.