logging: don't lose log messages during early startup
Previously, during early-startup (until configure_logging(config) is
called in run_electrum),
- the stderr log handler lost all log messages below warning level, and
- the file log handler lost all log messages regardless of log level

We now instead start buffering log messages in memory as soon as
electrum.logging is imported. The buffer is dumped into the
stderr and file log handlers when they are fully set up, and then
the buffer is discarded (and the temporary log handler is removed).

There is one issue with this implementation: the order of these early
messages is a bit scrambled... (for the stderr handler only)
This is because WARNING and ERROR messages are logged directly to stderr,
while lower level messages go into the buffer and are dumped later
(in _configure_stderr_verbosity). We could alternatively delay the WARNING+
messages and dump those too in _configure_stderr_verbosity, however
then they would not get shown if _configure_stderr_verbosity is never
called -- and this is likely the case when not using run_electrum.
(So the question is whether we want to be able to log warnings if someone
does `import electrum` in a python script.)