lnchannel: generate sweep_address on demand, to avoid some addr-reuse
I don't see a good reason to use the static_remotekey for sweeping (except maybe for the watchtower use case).

If there is a force-close, regardless if local or remote, the local payment_basepoint *will* appear on-chain
(if local-fclose, it will be used for the wpkh. if remote-fclose, it will be used in the CSV-locked output)
If we need to sweep anything, we should avoid reusing that pubkey.

This splits the "static" sweep_address property into two:
- a get_static_sweep_address_for_watchtower() method, which behaves the same as the property
  - see docstring for that method for reason of special-casing this
- a get_new_sweep_address() method used in all other cases, which asks the wallet for a fresh internal address
  - this avoids reusing the pubkey/address in some cases
    - but reuse can still happen in some cases, e.g. for a force-close with pending htlcs