qt: crash reporter: add "show details" btn with full raw report text
The msg box showing the crash report is not horizonatally scrollable and does not word-wrap,
so long lines (of a traceback) get truncated when displayed.
This adds a "Show details" btn, using built-in mechanisms, to expose the full report text that way.
(alternatively user can try to select text of whole dialog, and copy it to a text editor to inspect)

Note: the dialog does not allow vertical scrolling either. If the traceback is long,
that could prevent even the "show details" btn to be displayed on screen :/
(but selecting "all" text and copying it out should still work)